Add Moving Platform Prototype.
The moving platform moves between points with linear interpolation. Each
point may optionally set the speed of the platform, cause the platform
to stop at the point for a fixed amount of time, or cause the platform
to stop indefinitely (by setting the speed to zero, then setting the
speed to something else to make the platform move again). Each point
also sets the next point in the path.